All typings except one ParamSpec issue

+++ jishaku/shim/paginator_200.py
@@ -12,6 +12,7 @@ Paginator-related tools and interfaces for Jishaku.
 """
 
 import asyncio
+import typing
 
 import discord
 from discord import ui
@@ -57,8 +58,8 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
                 await interface.add_line("I'm still here!")
     """
 
-    def __init__(self, bot: commands.Bot, paginator: commands.Paginator, **kwargs):
-        if not isinstance(paginator, commands.Paginator):
+    def __init__(self, bot: commands.Bot, paginator: commands.Paginator, **kwargs: typing.Any):
+        if not isinstance(paginator, commands.Paginator):  # type: ignore
             raise TypeError('paginator must be a commands.Paginator instance')
 
         self._display_page = 0
@@ -75,10 +76,10 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
 
         self.sent_page_reactions = False
 
-        self.task: asyncio.Task = None
+        self.task: typing.Optional[asyncio.Task[None]] = None
         self.send_lock: asyncio.Event = asyncio.Event()
 
-        self.close_exception: Exception = None
+        self.close_exception: typing.Optional[BaseException] = None
 
         if self.page_size > self.max_page_size:
             raise ValueError(
@@ -96,9 +97,13 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
         # protected access has to be permitted here to not close the paginator's pages
 
         # pylint: disable=protected-access
-        paginator_pages = list(self.paginator._pages)
-        if len(self.paginator._current_page) > 1:
-            paginator_pages.append('\n'.join(self.paginator._current_page) + '\n' + (self.paginator.suffix or ''))
+        paginator_pages = list(self.paginator._pages)  # type: ignore
+        if len(self.paginator._current_page) > 1:  # type: ignore
+            paginator_pages.append(
+                '\n'.join(self.paginator._current_page)  # type: ignore
+                + '\n'
+                + (self.paginator.suffix or '')
+            )
         # pylint: enable=protected-access
 
         return paginator_pages
@@ -121,7 +126,7 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
         return self._display_page
 
     @display_page.setter
-    def display_page(self, value):
+    def display_page(self, value: int):
         """
         Sets the current page the paginator is on. Automatically pushes values inbounds.
         """
@@ -141,7 +146,7 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
         return self.paginator.max_size + len(f'\nPage {page_count}/{page_count}')
 
     @property
-    def send_kwargs(self) -> dict:
+    def send_kwargs(self) -> typing.Dict[str, typing.Any]:
         """
         A property that returns the kwargs forwarded to send/edit when updating the page.
 
@@ -159,13 +164,13 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
         """
 
         self.button_start.label = f"1 \u200b {self.emojis.start}"
-        self.button_previous.label = self.emojis.back
+        self.button_previous.label = str(self.emojis.back)
         self.button_current.label = str(self.display_page + 1)
-        self.button_next.label = self.emojis.forward
+        self.button_next.label = str(self.emojis.forward)
         self.button_last.label = f"{self.emojis.end} \u200b {self.page_count}"
         self.button_close.label = f"{self.emojis.close} \u200b Close paginator"
 
-    async def add_line(self, *args, **kwargs):
+    async def add_line(self, *args: typing.Any, **kwargs: typing.Any):
         """
         A proxy function that allows this PaginatorInterface to remain locked to the last page
         if it is already on it.
@@ -231,6 +236,12 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
         Waits on a loop for updates to the interface. This should not be called manually - it is handled by `send_to`.
         """
 
+        if not self.message:
+            raise RuntimeError("Message not set on PaginatorInterface")
+
+        if not self.bot.user:
+            raise RuntimeError("A PaginatorInterface cannot be started while the bot is offline")
+
         try:  # pylint: disable=too-many-nested-blocks
             while not self.bot.is_closed():
                 await asyncio.wait_for(self.send_lock_delayed(), timeout=self.timeout_length)
@@ -259,14 +270,14 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
             else:
                 await self.message.edit(view=None)
 
-    async def interaction_check(self, *args):
+    async def interaction_check(self, *args: typing.Any):
         """Check that determines whether this interaction should be honored"""
-        *_, interaction = args  # #149
+        *_, interaction = args  # type: ignore  #149
         interaction: discord.Interaction
         return not self.owner or interaction.user.id == self.owner.id
 
     @ui.button(label="1 \u200b \N{BLACK LEFT-POINTING DOUBLE TRIANGLE WITH VERTICAL BAR}", style=discord.ButtonStyle.secondary)
-    async def button_start(self, interaction: discord.Interaction, button: ui.Button):  # pylint: disable=unused-argument
+    async def button_start(self, interaction: discord.Interaction, button: ui.Button['PaginatorInterface']):  # pylint: disable=unused-argument
         """Button to send interface to first page"""
 
         self._display_page = 0
@@ -274,7 +285,7 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
         await interaction.response.edit_message(**self.send_kwargs)
 
     @ui.button(label="\N{BLACK LEFT-POINTING TRIANGLE}", style=discord.ButtonStyle.secondary)
-    async def button_previous(self, interaction: discord.Interaction, button: ui.Button):  # pylint: disable=unused-argument
+    async def button_previous(self, interaction: discord.Interaction, button: ui.Button['PaginatorInterface']):  # pylint: disable=unused-argument
         """Button to send interface to previous page"""
 
         self._display_page -= 1
@@ -282,14 +293,14 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
         await interaction.response.edit_message(**self.send_kwargs)
 
     @ui.button(label="1", style=discord.ButtonStyle.primary)
-    async def button_current(self, interaction: discord.Interaction, button: ui.Button):  # pylint: disable=unused-argument
+    async def button_current(self, interaction: discord.Interaction, button: ui.Button['PaginatorInterface']):  # pylint: disable=unused-argument
         """Button to refresh the interface"""
 
         self.update_view()
         await interaction.response.edit_message(**self.send_kwargs)
 
     @ui.button(label="\N{BLACK RIGHT-POINTING TRIANGLE}", style=discord.ButtonStyle.secondary)
-    async def button_next(self, interaction: discord.Interaction, button: ui.Button):  # pylint: disable=unused-argument
+    async def button_next(self, interaction: discord.Interaction, button: ui.Button['PaginatorInterface']):  # pylint: disable=unused-argument
         """Button to send interface to next page"""
 
         self._display_page += 1
@@ -297,7 +308,7 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
         await interaction.response.edit_message(**self.send_kwargs)
 
     @ui.button(label="\N{BLACK RIGHT-POINTING DOUBLE TRIANGLE WITH VERTICAL BAR} \u200b 1", style=discord.ButtonStyle.secondary)
-    async def button_last(self, interaction: discord.Interaction, button: ui.Button):  # pylint: disable=unused-argument
+    async def button_last(self, interaction: discord.Interaction, button: ui.Button['PaginatorInterface']):  # pylint: disable=unused-argument
         """Button to send interface to last page"""
 
         self._display_page = self.page_count - 1
@@ -307,9 +318,9 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
     class PageChangeModal(discord.ui.Modal, title="Go to page"):
         """Modal that prompts users for the page number to change to"""
 
-        page_number = discord.ui.TextInput(label="Page number", style=discord.TextStyle.short)
+        page_number: discord.ui.TextInput[discord.ui.Modal] = discord.ui.TextInput(label="Page number", style=discord.TextStyle.short)
 
-        def __init__(self, interface: 'PaginatorInterface', *args, **kwargs):
+        def __init__(self, interface: 'PaginatorInterface', *args: typing.Any, **kwargs: typing.Any):
             super().__init__(*args, timeout=interface.timeout_length, **kwargs)
             self.interface = interface
             self.page_number.label = f"Page number (1-{interface.page_count})"
@@ -318,6 +329,9 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
 
         async def on_submit(self, interaction: discord.Interaction):
             try:
+                if not self.page_number.value:
+                    raise ValueError("Page number not filled")
+
                 self.interface.display_page = int(self.page_number.value) - 1
             except ValueError:
                 await interaction.response.send_message(
@@ -329,20 +343,22 @@ class PaginatorInterface(ui.View):  # pylint: disable=too-many-instance-attribut
                 await interaction.response.edit_message(**self.interface.send_kwargs)
 
     @ui.button(label="\N{RIGHTWARDS ARROW WITH HOOK} \u200b Go to page", style=discord.ButtonStyle.primary)
-    async def button_goto(self, interaction: discord.Interaction, button: ui.Button):  # pylint: disable=unused-argument
+    async def button_goto(self, interaction: discord.Interaction, button: ui.Button['PaginatorInterface']):  # pylint: disable=unused-argument
         """Button to jump directly to a page"""
 
         await interaction.response.send_modal(self.PageChangeModal(self))
 
     @ui.button(label="\N{BLACK SQUARE FOR STOP} \u200b Close paginator", style=discord.ButtonStyle.danger)
-    async def button_close(self, interaction: discord.Interaction, button: ui.Button):  # pylint: disable=unused-argument
+    async def button_close(self, interaction: discord.Interaction, button: ui.Button['PaginatorInterface']):  # pylint: disable=unused-argument
         """Button to close the interface"""
 
         message = self.message
         self.message = None
-        self.task.cancel()
+        if self.task:
+            self.task.cancel()
         self.stop()
-        await message.delete()
+        if message:
+            await message.delete()
 
 
 class PaginatorEmbedInterface(PaginatorInterface):
@@ -350,12 +366,12 @@ class PaginatorEmbedInterface(PaginatorInterface):
     A subclass of :class:`PaginatorInterface` that encloses content in an Embed.
     """
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args: typing.Any, **kwargs: typing.Any):
         self._embed = kwargs.pop('embed', None) or discord.Embed()
         super().__init__(*args, **kwargs)
 
     @property
-    def send_kwargs(self) -> dict:
+    def send_kwargs(self) -> typing.Dict[str, typing.Any]:
         self._embed.description = self.pages[self.display_page]
         return {'embed': self._embed, 'view': self}
